Implemented gamepad emulation.
The timings are tweaked because it seems our emulator is too fast at
this point: the BIOS attempts to start a new request while the
previous one is still in progress. This will have to be fixed later as
we become more accurate.

Only the original digital controller is supported for now.

The frontend supports keyboard and controllers through SDL2 for input.