Support reading from CD-ROM XA mode 2 form 2 tracks
This seems buggy (since we don't retrieve the entire sector data) but
Spyro does just that during the intro sequence. Maybe it's caused by
an other emulation bug or maybe it's really supposed to do that. Needs
more investigating. At any rate a warning is displayed when this code
is triggered.